Move some things around
