+++ warehouse/metrics_tools/compute/test_app.py
@@ -4,8 +4,8 @@ from unittest.mock import MagicMock
 
 import pytest
 from fastapi.testclient import TestClient
+from metrics_tools.compute.app import app_factory, default_lifecycle
 from metrics_tools.compute.client import BaseWebsocketConnector, Client
-from metrics_tools.compute.server import app_factory, default_lifecycle
 from metrics_tools.compute.types import AppConfig
 from metrics_tools.definition import PeerMetricDependencyRef
 from metrics_tools.utils.logging import setup_module_logging
